Make S3 functions implement BucketStorage protocol
S3 code was used directly everywhere making it hard to use another
bucket storage service in place of s3 e.i. Google Cloud Storage.

This patch addresses the issue by defining a protocol `BucketStorage`
and using this protocol wherever s3 functions was called directly
before. It also defines an S3 record which implements the protocol such
that the behaviour is the same as before.